pyproject.toml
[project]
name = "miniature"
version = "2.0.0.dev"
description = "Python package for managing local git repositories and publishing packages with version tagging"
authors = [
    {name = "Sisung Kim", email = "sisung.kim1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "runpycli>=0.1.0",
    "pyshell",
    "packaging",
    "gitpython",
]

[project.scripts]
miniature = "miniature.cli:cli.app"
